refactor(#316): extract useStartupData and useCloudSync hooks from App.tsx
Removes another 69 lines from App.tsx (3009 → 2940):

1. useStartupData (web/src/hooks/useStartupData.ts)
   - Encapsulates the three mount-only startup effects: daily login reward
     peek, unclaimed developer gifts, and news unread count
   - Returns { dailyReward, setDailyReward, pendingGifts, setPendingGifts,
     newsUnreadCount, setNewsUnreadCount }

2. useCloudSync (web/src/hooks/useCloudSync.ts)
   - Encapsulates syncPromptedRef, the remote-save check on title screen,
     and the 5-minute periodic auto-upload interval
   - Returns { syncPrompt, clearSyncPrompt }; render handlers for
     "Load Remote" / "Keep Local" stay inline in App.tsx (2-3 lines each)
